added erode, dilate, apply-color-map with documentation and examples for each

+++ include/size.hpp
@@ -5,7 +5,7 @@
 extern "C" {
     Size2f* cv_create_Size2f(float width, float height);
     Size* cv_create_Size();
-    Size* cv_create_Size2(double width, double height);
+    Size* cv_create_Size2(int width, int height);
     Size* cv_Size_assignTo(Size* self, Size* other);
     Size* cv_Size_fromPoint(Point* p);
     double cv_Size_area(Size* self);
